softhsm2-proxy/web/requirements.txt
waitress==1.4.4
Flask==1.1.2
python-pkcs11==0.7.0